+++ ft_strncpy.c
@@ -2,10 +2,10 @@
 
 char	*ft_strncpy(char *dest, const char *src, size_t n)
 {
-	size_t i;
+	unsigned int i;
 
 	i = 0;
-	while (src[i] && i < n)
+	while (src[i] && (size_t)i < n)
 	{
 		dest[i] = src[i];
 		i += 1;
